requirements.txt
python-telegram-bot==20.7
python-decouple==3.8